feat: ec2 instance will now be started when there are running builds
Allows the EC2 Instance to correctly start if it is currently stopped,
but a build has become stuck in the 'running' state. This typically
occurs when a build begins in the 'dead zone' between an EC2 Instance
beginning the shutdown, but before the Drone Worker process is
terminated.

This patch should allow recovery from this state by starting the EC2
instance back up again to resume the job.